requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.0.0
ultralytics==8.3.243
ultralytics-thop==2.0.18
django
split-folders
tqdm

# Also download pip install torch==2.9.1 torchvision==0.24.1 torchaudio==2.9.1 --index-url https://download.pytorch.org/whl/cu126
